requirements.txt
PyYAML==6.0.2
aiohttp==3.12.13
azure-storage-blob==12.25.1
doi2pdf==0.1.3
fastapi[standard]==0.115.13
openai==1.91.0
pre-commit==4.2.0
pydantic==2.11.7
pytest==8.4.1
python-dotenv==1.1.0
requests==2.32.4
ruff==0.12.0
uvicorn==0.34.3
